Created a separate Eigenvalue Decomposition class and adjust AssembleScalarEigen to link into that class.  This will allow additional units to have access to the EVD solver.
