+++ src/EigenDecomposition.C
@@ -1,0 +1,275 @@
+/*------------------------------------------------------------------------*/
+/*  Copyright 2014 Sandia Corporation.                                    */
+/*  This software is released under the license detailed                  */
+/*  in the file, LICENSE, which is located in the top-level Nalu          */
+/*  directory structure                                                   */
+/*------------------------------------------------------------------------*/
+
+
+#include <EigenDecomposition.h>
+#include <NaluParsing.h>
+
+// basic c++
+#include <stdexcept>
+
+namespace sierra{
+namespace nalu{
+
+//==========================================================================
+// Class Definition
+//==========================================================================
+// EigenDecomposition -- Computes an eigenvalue decomposition and other
+//                    -- related functionality
+//==========================================================================
+//--------------------------------------------------------------------------
+//-------- constructor -----------------------------------------------------
+//--------------------------------------------------------------------------
+EigenDecomposition::EigenDecomposition() 
+{
+  // does nothing
+}
+
+//--------------------------------------------------------------------------
+//-------- destructor ------------------------------------------------------
+//--------------------------------------------------------------------------
+EigenDecomposition::~EigenDecomposition()
+{
+  // nothing to do
+}
+
+//--------------------------------------------------------------------------
+//-------- symmetric diagonalize (2D) --------------------------------------
+//--------------------------------------------------------------------------
+void
+EigenDecomposition::sym_diagonalize(
+  const double (&A)[2][2], double (&Q)[2][2], double (&D)[2][2])
+{
+  // Note that A must be symmetric here
+  const double trace = A[0][0] + A[1][1];
+  const double det = A[0][0]*A[1][1] - A[0][1]*A[1][0];
+
+  // calculate eigenvalues
+  D[0][0] = trace/2.0 + std::sqrt(trace*trace/4.0 - det);
+  D[1][1] = trace/2.0 - std::sqrt(trace*trace/4.0 - det);
+  D[0][1] = 0.0;
+  D[1][0] = 0.0;
+
+  //calculate first eigenvector
+  Q[0][0] = -A[0][1];
+  Q[1][0] =  A[0][0] - D[0][0];
+
+  double norm = std::sqrt(Q[0][0]*Q[0][0] + Q[1][0]*Q[1][0]);
+  Q[0][0] = Q[0][0]/norm;
+  Q[1][0] = Q[1][0]/norm;
+
+  // calculate second eigenvector
+  Q[0][1] = -A[1][1] + D[1][1];
+  Q[1][1] =  A[1][0];
+
+  norm = std::sqrt(Q[0][1]*Q[0][1] + Q[1][1]*Q[1][1]);
+  Q[0][1] = Q[0][1]/norm;
+  Q[1][1] = Q[1][1]/norm;
+
+  // special case when off diagonal entries were 0, we already had a diagonal matrix
+  Q[0][0] = (A[1][0] == 0.0) ? 1.0 : Q[0][0];
+  Q[0][1] = (A[1][0] == 0.0) ? 0.0 : Q[0][1];
+  Q[1][0] = (A[1][0] == 0.0) ? 0.0 : Q[1][0];
+  Q[1][1] = (A[1][0] == 0.0) ? 1.0 : Q[1][1];
+}
+
+//--------------------------------------------------------------------------
+//-------- reconstruct_matrix_from_decomposition ---------------------------
+//--------------------------------------------------------------------------
+void
+EigenDecomposition::reconstruct_matrix_from_decomposition(
+  const double (&D)[2][2], const double (&Q)[2][2], double (&A)[2][2])
+{
+  // A = Q*D*QT
+  double QT[2][2];
+  double B[2][2];
+
+  // compute QT
+  for (int i = 0; i < 2; i++) {
+    for (int j = 0; j < 2; j++) {
+      QT[j][i] = Q[i][j];
+    }
+  }
+  //mat-vec, B = Q*D
+  matrix_matrix_multiply(Q,D,B);
+
+  // mat-vec, A = (Q*D)*QT = B*QT
+  matrix_matrix_multiply(B,QT,A);
+}
+
+//--------------------------------------------------------------------------
+//-------- matrix_matrix_multiply ------------------------------------------
+//--------------------------------------------------------------------------
+void
+EigenDecomposition::matrix_matrix_multiply(
+  const double (&A)[2][2], const double (&B)[2][2], double (&C)[2][2])
+{
+  //C = A*B
+  for (int i = 0; i < 2; ++i) {
+    for (int j = 0; j < 2; ++j) {
+      double sum = 0;
+      for (int k = 0; k < 2; ++k) {
+        sum = sum + A[i][k] * B[k][j];
+      }
+      C[i][j] = sum;
+    }
+  }
+}
+
+//--------------------------------------------------------------------------
+//-------- symmetric diagonalize (3D) --------------------------------------
+//--------------------------------------------------------------------------
+void
+EigenDecomposition::sym_diagonalize(
+  const double (&A)[3][3], double (&Q)[3][3], double (&D)[3][3])
+{
+  /*
+    obtained from: 
+    http://stackoverflow.com/questions/4372224/
+    fast-method-for-computing-3x3-symmetric-matrix-spectral-decomposition
+
+    A must be a symmetric matrix.
+    returns Q and D such that 
+    Diagonal matrix D = QT * A * Q;  and  A = Q*D*QT
+  */
+
+  const int maxsteps=24;
+  int k0, k1, k2;
+  double o[3], m[3];
+  double q [4] = {0.0,0.0,0.0,1.0};
+  double jr[4];
+  double sqw, sqx, sqy, sqz;
+  double tmp1, tmp2, mq;
+  double AQ[3][3];
+  double thet, sgn, t, c;
+  for(int i=0;i < maxsteps;++i) {
+    // quat to matrix
+    sqx      = q[0]*q[0];
+    sqy      = q[1]*q[1];
+    sqz      = q[2]*q[2];
+    sqw      = q[3]*q[3];
+    Q[0][0]  = ( sqx - sqy - sqz + sqw);
+    Q[1][1]  = (-sqx + sqy - sqz + sqw);
+    Q[2][2]  = (-sqx - sqy + sqz + sqw);
+    tmp1     = q[0]*q[1];
+    tmp2     = q[2]*q[3];
+    Q[1][0]  = 2.0 * (tmp1 + tmp2);
+    Q[0][1]  = 2.0 * (tmp1 - tmp2);
+    tmp1     = q[0]*q[2];
+    tmp2     = q[1]*q[3];
+    Q[2][0]  = 2.0 * (tmp1 - tmp2);
+    Q[0][2]  = 2.0 * (tmp1 + tmp2);
+    tmp1     = q[1]*q[2];
+    tmp2     = q[0]*q[3];
+    Q[2][1]  = 2.0 * (tmp1 + tmp2);
+    Q[1][2]  = 2.0 * (tmp1 - tmp2);
+
+    // AQ = A * Q
+    AQ[0][0] = Q[0][0]*A[0][0]+Q[1][0]*A[0][1]+Q[2][0]*A[0][2];
+    AQ[0][1] = Q[0][1]*A[0][0]+Q[1][1]*A[0][1]+Q[2][1]*A[0][2];
+    AQ[0][2] = Q[0][2]*A[0][0]+Q[1][2]*A[0][1]+Q[2][2]*A[0][2];
+    AQ[1][0] = Q[0][0]*A[0][1]+Q[1][0]*A[1][1]+Q[2][0]*A[1][2];
+    AQ[1][1] = Q[0][1]*A[0][1]+Q[1][1]*A[1][1]+Q[2][1]*A[1][2];
+    AQ[1][2] = Q[0][2]*A[0][1]+Q[1][2]*A[1][1]+Q[2][2]*A[1][2];
+    AQ[2][0] = Q[0][0]*A[0][2]+Q[1][0]*A[1][2]+Q[2][0]*A[2][2];
+    AQ[2][1] = Q[0][1]*A[0][2]+Q[1][1]*A[1][2]+Q[2][1]*A[2][2];
+    AQ[2][2] = Q[0][2]*A[0][2]+Q[1][2]*A[1][2]+Q[2][2]*A[2][2];
+    // D = Qt * AQ
+    D[0][0] = AQ[0][0]*Q[0][0]+AQ[1][0]*Q[1][0]+AQ[2][0]*Q[2][0];
+    D[0][1] = AQ[0][0]*Q[0][1]+AQ[1][0]*Q[1][1]+AQ[2][0]*Q[2][1];
+    D[0][2] = AQ[0][0]*Q[0][2]+AQ[1][0]*Q[1][2]+AQ[2][0]*Q[2][2];
+    D[1][0] = AQ[0][1]*Q[0][0]+AQ[1][1]*Q[1][0]+AQ[2][1]*Q[2][0];
+    D[1][1] = AQ[0][1]*Q[0][1]+AQ[1][1]*Q[1][1]+AQ[2][1]*Q[2][1];
+    D[1][2] = AQ[0][1]*Q[0][2]+AQ[1][1]*Q[1][2]+AQ[2][1]*Q[2][2];
+    D[2][0] = AQ[0][2]*Q[0][0]+AQ[1][2]*Q[1][0]+AQ[2][2]*Q[2][0];
+    D[2][1] = AQ[0][2]*Q[0][1]+AQ[1][2]*Q[1][1]+AQ[2][2]*Q[2][1];
+    D[2][2] = AQ[0][2]*Q[0][2]+AQ[1][2]*Q[1][2]+AQ[2][2]*Q[2][2];
+    o[0]    = D[1][2];
+    o[1]    = D[0][2];
+    o[2]    = D[0][1];
+    m[0]    = std::abs(o[0]);
+    m[1]    = std::abs(o[1]);
+    m[2]    = std::abs(o[2]);
+
+    k0      = (m[0] > m[1] && m[0] > m[2])?0: (m[1] > m[2])? 1 : 2; // index of largest element of offdiag
+    k1      = (k0+1)%3;
+    k2      = (k0+2)%3;
+    if (o[k0]==0.0) {
+      break;  // diagonal already
+    }
+    thet    = (D[k2][k2]-D[k1][k1])/(2.0*o[k0]);
+    sgn     = (thet > 0.0)?1.0:-1.0;
+    thet   *= sgn; // make it positive
+    t       = sgn /(thet +((thet < 1.E6)? std::sqrt(thet*thet+1.0):thet)) ; // sign(T)/(|T|+sqrt(T^2+1))
+    c       = 1.0/std::sqrt(t*t+1.0); //  c= 1/(t^2+1) , t=s/c 
+    if(c==1.0) {
+      break;  // no room for improvement - reached machine precision.
+    }
+    jr[0 ]  = jr[1] = jr[2] = jr[3] = 0.0;
+    jr[k0]  = sgn*std::sqrt((1.0-c)/2.0);  // using 1/2 angle identity sin(a/2) = std::sqrt((1-cos(a))/2)  
+    jr[k0] *= -1.0; // since our quat-to-matrix convention was for v*M instead of M*v
+    jr[3 ]  = std::sqrt(1.0f - jr[k0] * jr[k0]);
+    if(jr[3]==1.0) {
+      break; // reached limits of floating point precision
+    }
+    q[0]    = (q[3]*jr[0] + q[0]*jr[3] + q[1]*jr[2] - q[2]*jr[1]);
+    q[1]    = (q[3]*jr[1] - q[0]*jr[2] + q[1]*jr[3] + q[2]*jr[0]);
+    q[2]    = (q[3]*jr[2] + q[0]*jr[1] - q[1]*jr[0] + q[2]*jr[3]);
+    q[3]    = (q[3]*jr[3] - q[0]*jr[0] - q[1]*jr[1] - q[2]*jr[2]);
+    mq      = std::sqrt(q[0] * q[0] + q[1] * q[1] + q[2] * q[2] + q[3] * q[3]);
+    q[0]   /= mq;
+    q[1]   /= mq;
+    q[2]   /= mq;
+    q[3]   /= mq;
+  }
+}
+
+//--------------------------------------------------------------------------
+//-------- reconstruct_matrix_from_decomposition ---------------------------
+//--------------------------------------------------------------------------
+void
+EigenDecomposition::reconstruct_matrix_from_decomposition(
+  const double (&D)[3][3], const double (&Q)[3][3], double (&A)[3][3])
+{
+  // A = Q*D*QT
+  double QT[3][3];
+  double B[3][3];
+
+  // compute QT
+  for (int i = 0; i < 3; i++) {
+    for (int j = 0; j < 3; j++) {
+      QT[j][i] = Q[i][j];
+    }
+  }
+  //mat-vec, B = Q*D
+  matrix_matrix_multiply(Q,D,B);
+
+  // mat-vec, A = (Q*D)*QT = B*QT
+  matrix_matrix_multiply(B,QT,A);
+}
+
+//--------------------------------------------------------------------------
+//-------- matrix_matrix_multiply ------------------------------------------
+//--------------------------------------------------------------------------
+void
+EigenDecomposition::matrix_matrix_multiply(
+  const double (&A)[3][3], const double (&B)[3][3], double (&C)[3][3])
+{
+  //C = A*B
+  for (int i = 0; i < 3; ++i) {
+    for (int j = 0; j < 3; ++j) {
+      double sum = 0;
+      for (int k = 0; k < 3; ++k) {
+        sum = sum + A[i][k] * B[k][j];
+      }
+      C[i][j] = sum;
+    }
+  }
+}
+
+} // namespace nalu
+} // namespace Sierra
